refactor: changed directory structure
