+++ Log4ShellDetector/Log4ShellDetector.py
