--- tests.mk
@@ -1,3 +1,0 @@
-
-tests:
-	pytest
